+++ src/TestSuite/MenuHelperTestCase.php
@@ -18,16 +18,53 @@ declare(strict_types=1);
 namespace MeCms\TestSuite;
 
 use Authentication\Identity;
+use MeCms\View\Helper\MenuHelper;
 use MeTools\TestSuite\HelperTestCase;
 use MeTools\View\Helper\HtmlHelper;
 
 /**
  * Abstract class for test `MenuHelper` classes
- * @property \MeCms\View\Helper\MenuHelper&\PHPUnit\Framework\MockObject\MockObject $Helper
- * @property class-string<\Cake\View\Helper> $originClassName
+ * @property class-string<\MeCms\View\Helper\MenuHelper> $originClassName
  */
 abstract class MenuHelperTestCase extends HelperTestCase
 {
+    /**
+     * @var \MeCms\View\Helper\MenuHelper&\PHPUnit\Framework\MockObject\MockObject
+     */
+    protected MenuHelper $Helper;
+
+    /**
+     * Called before every test method
+     * @return void
+     * @throws \ErrorException
+     */
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        if (empty($this->Helper)) {
+            $methodsToMock = get_child_methods($this->originClassName);
+            /** @var \MeCms\View\Helper\MenuHelper&\PHPUnit\Framework\MockObject\MockObject $Helper */
+            $Helper = $this->getMockForHelper($this->originClassName, $methodsToMock);
+
+            $OriginalHelper = new $this->originClassName($Helper->getView());
+            $HtmlHelper = new HtmlHelper($Helper->getView());
+
+            //Each method returns its original value, but links (the first result value) are already built and returned as HTML string
+            foreach ($methodsToMock as $method) {
+                $Helper->method($method)->willReturnCallback(function () use ($OriginalHelper, $HtmlHelper, $method) {
+                    $result = $OriginalHelper->$method();
+
+                    return empty($result) ? [] : [implode('', array_map(fn(array $link): string => $HtmlHelper->link(...$link), $result[0]))] + $result;
+                });
+            }
+
+            $this->Helper = $Helper;
+        }
+
+        $this->setIdentity(['group' => ['name' => 'user']]);
+    }
+
     /**
      * Internal method to set the identity for the current helper
      * @param array $data Identity data
@@ -37,47 +74,8 @@ abstract class MenuHelperTestCase extends HelperTestCase
     {
         $Request = $this->Helper->getView()->getRequest()->withAttribute('identity', new Identity($data));
         $this->Helper->getView()->setRequest($Request);
-        $this->Helper->Identity->initialize([]);
-    }
-
-    /**
-     * Get magic method.
-     *
-     * It provides access to the cached properties of the test.
-     * @param string $name Property name
-     * @return mixed
-     * @throws \ReflectionException
-     * @throws \Throwable
-     */
-    public function __get(string $name)
-    {
-        //Rewrites the parent method
-        if ($name === 'Helper') {
-            if (empty($this->_cache['Helper'])) {
-                $methods = get_child_methods($this->originClassName);
-                $Helper = $this->getMockForHelper($this->originClassName, $methods);
-                $OriginalHelper = new $this->originClassName($Helper->getView());
-                $HtmlHelper = new HtmlHelper($Helper->getView());
-
-                //Each method returns its original value, but links are already built and returned as HTML string
-                foreach ($methods as $method) {
-                    $Helper->method($method)->willReturnCallback(function () use ($OriginalHelper, $HtmlHelper, $method): array {
-                        $result = $OriginalHelper->$method();
-
-                        if (!empty($result[0])) {
-                            $result[0] = implode('', array_map(fn(array $link): string => $HtmlHelper->link(...$link), $result[0]));
-                        }
-
-                        return $result;
-                    });
-                }
-
-                $this->_cache['Helper'] = $Helper;
-            }
-
-            return $this->_cache['Helper'];
+        if ($this->Helper->Identity) {
+            $this->Helper->Identity->initialize([]);
         }
-
-        return parent::__get($name);
     }
 }